+++ soundpool_web/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 1.0.3
+* Support repeating sound in `play()`
+* Fallback to `stop()` when `pause()` was requested
 ## 1.0.2
 * Fix iOS building issues
 ## 1.0.1